requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.10.3
requests==2.22.0
